+++ lib/girl/tund.rb
@@ -8,14 +8,8 @@ module Girl
     PACK_SIZE = 1460
     CHUNK_SIZE = PACK_SIZE * 1000
 
-    def initialize( roomd_port = 9090, resend_port = 9091, dest_chunk_dir = '/tmp', tund_chunk_dir = '/tmp', resend_times = 20 )
-      # infos，存取sock信息，根据角色，sock => {}：
-      #
-      # {
-      #   id: '',
-      #   role: :roomd,
-      #   mon: mon
-      # }
+    def initialize( roomd_port = 9090, dest_chunk_dir = '/tmp', tund_chunk_dir = '/tmp', resend_times = 20 )
+      # infos 存取sock信息，根据角色，sock => {}
       #
       # {
       #   id: '',
@@ -44,6 +38,8 @@ module Girl
       #   chunks: [],
       #   chunk_seed: 0,
       #   ctl2: '',
+      #   mems_size: 0,
+      #   resends: [],
       #   rcur: 0,
       #   err_from_dest: 1,
       #   last_from_dest: 1024,
@@ -51,29 +47,35 @@ module Girl
       #   tun_addr: addrinfo
       # }
       #
-      # id:               用于文件缓存文件名，跨进程唯一
-      # role:             角色，:roomd / :dest / :tund
-      # mon:              NIO::Monitor
-      # close_by:         关闭标记，流量传完后关闭sock
-      # wbuff:            写缓存
-      # cache:            写的时候：先取cache，cache为空，取一个chunk放进cache，chunks也为空，取wbuff。
-      # chunk_dir:        文件缓存目录
-      # chunks:           文件缓存
-      # chunk_seed:       文件自增序号
-      # pieces:           跳号缓存
-      # last_from_source: 对面source的最后一个包序号
-      # pcur:             打包光标，打包dest流量，最后一个进tund写缓存的包号。
-      # tund:             对应的tund
-      # room_addr:        ctl1的来源地址
-      # ctl2:             ctl msg 2
-      # rcur:             读光标，读tun流量，最后一个进dest写缓存的包号。（跳号包放pieces）
-      # err_from_dest:    1 eof / 2 rst
-      # last_from_dest:   dest读到异常时的打包光标
-      # dest:             对应的dest
-      # tun_addr:         另一头地址
+      # id               用于文件缓存文件名，跨进程唯一
+      # role             角色，:roomd / :dest / :tund
+      # mon              NIO::Monitor
+      # close_by         关闭标记，流量传完后关闭sock
+      # wbuff            写缓存
+      # cache            写的时候：先取cache，cache为空，取一个chunk放进cache，chunks也为空，取wbuff。
+      # chunk_dir        文件缓存目录
+      # chunks           文件缓存
+      # chunk_seed       文件自增序号
+      # pieces           跳号缓存
+      # last_from_source 对面source的最后一个包序号
+      # pcur             打包光标，打包dest流量，最后一个进tund写缓存的包号。
+      # tund             对应的tund
+      # room_addr        ctl1的来源地址
+      # ctl2             ctl msg 2
+      # mems_size        写后计数
+      # resends          重传缓存
+      # rcur             读光标，读tun流量，最后一个进dest写缓存的包号。（跳号包放pieces）
+      # err_from_dest    1 eof / 2 rst
+      # last_from_dest   dest读到异常时的打包光标
+      # dest             对应的dest
+      # tun_addr         另一头地址
       @infos = {}
       # 写后缓存
+      # {
+      #   [ sock, pack_id ] => [ '', Time.new, 0 ]
+      # }
       @memories = {}
+      @mutex = Mutex.new
       @selector = NIO::Selector.new
       @dest_chunk_dir = dest_chunk_dir
       @tund_chunk_dir = tund_chunk_dir
@@ -84,28 +86,78 @@ module Girl
       @roomd.bind( Socket.pack_sockaddr_in( roomd_port, '0.0.0.0' ) )
       puts "roomd listening on #{ roomd_port }"
 
-      resend = Socket.new( Socket::AF_INET, Socket::SOCK_DGRAM, 0 )
-      resend.setsockopt( Socket::SOL_SOCKET, Socket::SO_REUSEPORT, 1 )
-      resend.bind( Socket.sockaddr_in( resend_port, '127.0.0.1' ) )
-      puts "resend listening on #{ resend_port }"
-
       roomd_mon = @selector.register( @roomd, :r )
       @infos[ @roomd ] = {
         id: [ Process.pid, @roomd.object_id ].join( '-' ),
         role: :roomd,
         mon: roomd_mon
       }
-
-      resend_mon = @selector.register( resend, :r )
-      @infos[ resend ] = {
-        id: [ Process.pid, resend.object_id ].join( '-' ),
-        role: :resend,
-        mon: resend_mon
-      }
     end
 
     def looping
       puts 'looping'
+
+      # 一秒重传
+      Thread.new do
+        loop do
+          memory = @memories.first
+
+          if memory
+            now = Time.new
+            key, mem = memory
+            data, mem_at, times = mem
+
+            if now - mem_at > 1
+              sock, pack_id = key
+
+              @mutex.synchronize do
+                @memories.delete( key )
+
+                if sock.closed?
+                  next
+                end
+
+                # 超过重传次数关闭通道
+                if times > @resend_times
+                  puts 'resend too many times'
+                  close_tund( sock )
+                  next
+                end
+
+                info = @infos[ sock ]
+                pack_id = data[ 0, 4 ].unpack( 'N' ).first
+
+                begin
+                  sock.sendmsg( data, 0, info[ :tun_addr ] )
+                rescue Errno::ENETUNREACH, IOError => e
+                  puts "resend #{ e.class }"
+                  close_tund( sock )
+                  next
+                end
+
+                @memories[ [ sock, pack_id ] ] = [ data, now, times + 1 ]
+
+                # 如果是最后一个包，发送close1
+                if info[ :last_from_dest ] && ( pack_id == info[ :last_from_dest ] )
+                  pack = [ 0, 5, info[ :err_from_dest ], pack_id ].pack( 'NCCN' )
+
+                  begin
+                    sock.sendmsg( pack, 0, info[ :tun_addr ] )
+                  rescue Errno::ENETUNREACH, IOError => e
+                    puts "resend close1 #{ e.class }"
+                    close_tund( sock )
+                    next
+                  end
+                end
+              end
+
+              next
+            end
+          end
+
+          sleep 0.01
+        end
+      end
 
       loop do
         @selector.select do | mon |
@@ -113,15 +165,6 @@ module Girl
 
           if sock.closed?
             puts 'sock already closed'
-
-            if mon.readable?
-              puts 'skip read'
-            end
-
-            if mon.writable?
-              puts 'skip write'
-            end
-
             next
           end
 
@@ -197,50 +240,14 @@ module Girl
                 chunk_seed: 0,
                 room_addr: addrinfo,
                 ctl2: ctl2,
+                mems_size: 0,
+                resends: [],
                 rcur: 0,
                 err_from_dest: nil,
                 last_from_dest: nil,
                 dest: dest,
                 tun_addr: nil
               }
-
-              @memories[ tund ] = {}
-            when :resend
-              # 重传
-              data, addrinfo, rflags, *controls = sock.recvmsg
-              now = Time.new
-
-              @memories.each do | tund, mems |
-                tund_info = @infos[ tund ]
-
-                # 一秒重传
-                mems.select{ | pack_id, mem | now - mem[ 1 ] >= 1 }.each do | pack_id, mem |
-                  mem_data, mem_at, times = mem
-
-                  # 重传超过x次关闭通道
-                  if times >= @resend_times
-                    puts 'resend too many times'
-                    close_tund( tund )
-                    break
-                  end
-
-                  begin
-                    tund.sendmsg( mem_data, 0, tund_info[ :tun_addr ] )
-                  rescue Errno::ENETUNREACH => e
-                    puts "send to tun #{ e.class }"
-                    close_tund( tund )
-                    break
-                  end
-
-                  # 重传close1
-                  if tund_info[ :last_from_dest ] && ( pack_id == tund_info[ :last_from_dest ] )
-                    ctlmsg = [ 5, tund_info[ :err_from_dest ], pack_id ].pack( 'CCN' )
-                    write_buff2( tund_info, pack_ctlmsg( ctlmsg ) )
-                  end
-
-                  @memories[ tund ][ pack_id ] = [ mem_data, now, times + 1 ]
-                end
-              end
             when :dest
               # 读dest，放进tund的写缓存
               begin
@@ -258,7 +265,8 @@ module Girl
                 data = @hex.swap( data )
               end
 
-              data = "#{ [ 4 + data.bytesize ].pack( 'n' ) }#{ [ pack_id ].pack( 'N' ) }#{ data }"
+              # 流量长度，包号，流量
+              data = [ [ data.bytesize, pack_id ].pack( 'nN' ), data ].join
               tund_info = @infos[ info[ :tund ] ]
 
               if tund_info[ :tun_addr ]
@@ -288,7 +296,8 @@ module Girl
                 when 4
                   # 4 confirm a pack -> N: pack_id
                   confirm_id = data[ 5, 4 ].unpack( 'N' ).first
-                  @memories[ sock ].delete( confirm_id )
+                  @memories.delete( [ sock, confirm_id ] )
+                  info[ :mems_size ] -= 1
                   info[ :mon ].add_interest( :w )
                 when 6
                   # 6 source close2
@@ -345,7 +354,7 @@ module Girl
           if mon.writable?
             case info[ :role ]
             when :dest
-              data = read_buff( info )
+              data, from = read_buff( info )
 
               if data.empty?
                 tund_info = @infos[ info[ :tund ] ]
@@ -378,17 +387,17 @@ module Girl
 
               data = data[ written..-1 ]
 
-              if info[ :cache ].empty?
-                info[ :wbuff ] = data
-              else
+              if from == :cache
                 info[ :cache ] = data
+              elsif from == :wbuff
+                info[ :wbuff ] = data
               end
             when :tund
               # tund的第一次写兴趣，roomd返回tund端口号给申请者
               unless info[ :tun_addr ]
                 begin
-                  @roomd.sendmsg( "#{ [ 0 ].pack( 'N' ) }#{ info[ :ctl2 ] }", 0, info[ :room_addr ] )
-                rescue Errno::ENETUNREACH => e
+                  @roomd.sendmsg( [ [ 0 ].pack( 'N' ), info[ :ctl2 ] ].join, 0, info[ :room_addr ] )
+                rescue Errno::ENETUNREACH, IOError => e
                   puts "send to room #{ e.class }"
                   close_tund( sock )
                   next
@@ -398,17 +407,10 @@ module Girl
                 next
               end
 
-              # 写后缓存超过1000，停止写
-              if @memories[ sock ].size > 1000
-                mon.remove_interest( :w )
-                next
-              end
-
-              # 取写前缓存
-              data = read_buff( info )
+              # 取写缓存
+              data, from = read_buff( info )
 
               if data.empty?
-                # 有关闭标记，关闭tund
                 if info[ :close_by ]
                   close_sock( sock )
                 else
@@ -419,26 +421,33 @@ module Girl
               end
 
               len, pack_id = data[ 0, 6 ].unpack( 'nN' )
-              pack = data[ 2, len ]
+              pack = data[ 2, ( 4 + len ) ]
 
               begin
                 sock.sendmsg( pack, 0, info[ :tun_addr ] )
-              rescue Errno::ENETUNREACH => e
+              rescue Errno::ENETUNREACH, IOError => e
                 puts "send to tun #{ e.class }"
                 close_tund( sock )
                 next
               end
 
               if pack_id > 0
-                @memories[ sock ][ pack_id ] = [ pack, Time.new, 0 ]
+                @memories[ [ sock, pack_id ] ] = [ pack, Time.new, 0 ]
+                info[ :mems_size ] += 1
+
+                # 如果是最后一个包，发送close1
+                if info[ :last_from_dest ] && ( pack_id == info[ :last_from_dest ] )
+                  ctlmsg = [ 5, info[ :err_from_dest ], pack_id ].pack( 'CCN' )
+                  write_buff2( info, pack_ctlmsg( ctlmsg ) )
+                end
               end
 
-              data = data[ ( 2 + len )..-1 ]
+              data = data[ ( 6 + len )..-1 ]
 
-              if info[ :cache ].empty?
-                info[ :wbuff ] = data
-              else
+              if from == :cache
                 info[ :cache ] = data
+              elsif from == :wbuff
+                info[ :wbuff ] = data
               end
             end
           end
@@ -470,10 +479,15 @@ module Girl
       # 先取cache
       # cache为空，取一个chunk放进cache
       # chunks也为空，取wbuff
-      data = info[ :cache ]
+      data, from = info[ :cache ], :cache
 
       if data.empty?
         if info[ :chunks ].any?
+          # tund写后超过1000，限制内存，暂不读入写前文件
+          if info[ :role ] == :tund && info[ :mems_size ] > 1000
+            return [ data, from ]
+          end
+
           path = File.join( info[ :chunk_dir ], info[ :chunks ].shift )
           data = info[ :cache ] = IO.binread( path )
 
@@ -482,15 +496,16 @@ module Girl
           rescue Errno::ENOENT
           end
         else
-          data = info[ :wbuff ]
+          data, from = info[ :wbuff ], :wbuff
         end
       end
 
-      return data
+      [ data, from ]
     end
 
     def pack_ctlmsg( data )
-      "#{ [ data.bytesize + 4, 0 ].pack( 'nN' ) }#{ data }"
+      # 流量长度，包号，流量
+      [ [ data.bytesize, 0 ].pack( 'nN' ), data ].join
     end
 
     def close_dest( sock, e )
@@ -512,7 +527,7 @@ module Girl
     def close_tund( sock )
       info = close_sock( sock )
 
-      unless info[ :dest ].closed?
+      if info && !info[ :dest ].closed?
         info[ :dest ].setsockopt( Socket::SOL_SOCKET, Socket::SO_LINGER, [ 1, 0 ].pack( 'ii' ) )
         close_sock( info[ :dest ] )
       end
@@ -523,14 +538,17 @@ module Girl
       @selector.deregister( sock )
       @memories.delete( sock )
       info = @infos.delete( sock )
-      info[ :chunks ].each do | filename |
-        begin
-          File.delete( File.join( info[ :chunk_dir ], filename ) )
-        rescue Errno::ENOENT
+
+      if info
+        info[ :chunks ].each do | filename |
+          begin
+            File.delete( File.join( info[ :chunk_dir ], filename ) )
+          rescue Errno::ENOENT
+          end
         end
       end
 
-      return info
+      info
     end
   end
 end
